Update gatsby-node.js
I added 'createSchemaCustomization' in my gatsby-node.js because I was suddenly not able to run gatsby in development mode and gatsby-cli suggested it. 

Worked for me. But now there was another problem. 

Apparently the build also ran error-free. But my first record in allGhostPost was the fake Node. I have these Nodes in my gatsby-node.js allready defined and after I commented the 'createTemporaryFakeNodes()' in 'onPreExtractQueries()' out, it finally worked.